requirements.txt
requests==2.22.0
python-dotenv==0.12.0
lxml==4.5.0
